pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
description = "HTTP Frontend as a django app for argus-server using HTMx"
name = "argus-htmx-frontend"
readme = "README.rst"
authors = [{name = "Hanne Moa", email = "hanne.moa@sikt.no"}]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Programming Language :: Python :: 3 :: Only",
    "Environment :: Web Environment",
]
dynamic = ["version"]
dependencies = [
    "argus-server>=1.20.1",
    "django-htmx",
    "django-widget-tweaks==1.5.0",
    "django-template-partials",
    # The next is for python_version < 3.12 but we simplify code by always needing it
    "importlib_resources>=5.12",
]

[project.urls]
Github = "https://github.com/Uninett/argus-htmx-frontend"

[project.optional-dependencies]
docs = ["sphinx"]

[tool.flit.module]
name = "argus_htmx"
